requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
openpyxl==3.1.2
pydantic==2.5.0
pydantic-core==2.14.3
matplotlib==3.8.2
seaborn==0.13.0
requests-ratelimiter==0.4.0
streamlit==1.29.0
beautifulsoup4==4.12.2
concurrent-log-handler==0.9.24 